+++ app/features/write-reviews/sections/ReviewRightSection/PopularFeedSubSection.tsx
@@ -1,8 +1,7 @@
-import styled from "@emotion/styled"
 import { useTranslation } from "react-i18next"
 
 import type { GETReviewsResponse } from "@/api/reviews"
-import ReviewBlock from "@/common/components/blocks/ReviewBlock"
+import ReviewBlock from "@/common/components/reviews/ReviewBlock"
 import FlexWrapper from "@/common/primitives/FlexWrapper"
 import Typography from "@/common/primitives/Typography"
 
@@ -10,13 +9,6 @@ interface PopularFeedSubSectionProps {
     reviews: GETReviewsResponse
     likeReview: (reviewId: number) => void
 }
-
-const ReviewWrapper = styled(FlexWrapper)`
-    padding: 8px 6px;
-    border: 1px ${({ theme }) => theme.colors.Background.Block.dark} solid;
-    background-color: ${({ theme }) => theme.colors.Background.Block.default};
-    border-radius: 6px;
-`
 
 function PopularFeedSubSection({ reviews, likeReview }: PopularFeedSubSectionProps) {
     const { t } = useTranslation()
@@ -29,16 +21,12 @@ function PopularFeedSubSection({ reviews, likeReview }: PopularFeedSubSectionPro
                 </Typography>
             </FlexWrapper>
             <FlexWrapper direction="column" align="stretch" gap={12}>
-                {reviews.reviews.map((review, idx) => (
-                    <ReviewWrapper
-                        direction="column"
-                        align="stretch"
-                        gap={0}
-                        padding="8px 10px"
-                        key={idx}
-                    >
-                        <ReviewBlock review={review} likeReview={likeReview} />
-                    </ReviewWrapper>
+                {reviews.reviews.map((review) => (
+                    <ReviewBlock
+                        review={review}
+                        likeReview={likeReview}
+                        key={review.id}
+                    />
                 ))}
             </FlexWrapper>
         </FlexWrapper>
