+++ app/features/write-reviews/sections/ReviewRightSection/ReviewFeedSubSection.tsx
@@ -5,7 +5,7 @@ import { Trans, useTranslation } from "react-i18next"
 
 import type { GETReviewsResponse } from "@/api/reviews"
 import ScrollableDropdown from "@/common/components/ScrollableDropdown"
-import ReviewBlock from "@/common/components/blocks/ReviewBlock"
+import ReviewBlock from "@/common/components/reviews/ReviewBlock"
 import { SemesterEnum, semesterToString } from "@/common/enum/semesterEnum"
 import FlexWrapper from "@/common/primitives/FlexWrapper"
 import Typography from "@/common/primitives/Typography"
@@ -18,13 +18,6 @@ const DropDownWrapper = styled(FlexWrapper)`
     flex: 1 0 1;
     width: 225px;
     height: 36px;
-`
-
-const ReviewWrapper = styled(FlexWrapper)`
-    padding: 8px 6px;
-    border: 1px ${({ theme }) => theme.colors.Background.Block.dark} solid;
-    background-color: ${({ theme }) => theme.colors.Background.Block.default};
-    border-radius: 6px;
 `
 
 const options = ["2023", "2022", "2021", "2020", "2019", "2018", "2017"]
@@ -67,16 +60,12 @@ function ReviewFeedSubSection({ reviews, likeReview }: ReviewFeedSubSectionProps
                 </FlexWrapper>
             </FlexWrapper>
             <FlexWrapper direction="column" align="stretch" gap={12}>
-                {reviews.reviews.map((review, idx) => (
-                    <ReviewWrapper
-                        direction="column"
-                        align="stretch"
-                        gap={0}
-                        padding="8px 10px"
-                        key={idx}
-                    >
-                        <ReviewBlock review={review} likeReview={likeReview} />
-                    </ReviewWrapper>
+                {reviews.reviews.map((review) => (
+                    <ReviewBlock
+                        review={review}
+                        likeReview={likeReview}
+                        key={review.id}
+                    />
                 ))}
             </FlexWrapper>
         </FlexWrapper>
